HAJ-1: update company lite

+++ shj-admin-portal-services/src/main/java/com/elm/shj/admin/portal/services/data/huic/HuicCompany.java
@@ -53,7 +53,7 @@ public class HuicCompany implements Serializable {
     private Long crNumber;
 
     @CompanyTypeCode
-    private Integer companyTypeCode;
+    private Long companyTypeCode;
 
     @RitualTypeCode
     private Long ritualTypeCode;
