+++ shj-admin-portal-services/src/main/java/com/elm/shj/admin/portal/services/data/huic/HuicPlannedPackage.java
@@ -49,7 +49,7 @@ public class HuicPlannedPackage implements Serializable {
     @WithPackageType
     private Long packageTypeCode;
     @CompanyTypeCode
-    private Integer companyTypeCode;
+    private Long companyTypeCode;
     @EstablishmentCode
     private Long establishmentId;
